Lots of changes/additions: * Add libpcsound codeblocks project * Use libraries that are built and project dependencies inside main   chocolate-doom/setup project and remove source files. * libraries are now built in the "lib" top level directory * Fix some warnings
Subversion-branch: /trunk/chocolate-doom
Subversion-revision: 852

+++ src/i_pcsound.c
@@ -84,7 +84,7 @@ void PCSCallbackFunc(int *duration, int *freq)
         // for a full discussion of this.
         // Check we don't overflow the frequency table.
 
-        if (tone < NUM_FREQUENCIES)
+        if (tone < (int)NUM_FREQUENCIES)
         {
             *freq = (int) frequencies[tone];
         }
